requirements.txt
flask==3.0.0
pandas==2.1.4
numpy==1.24.3
yfinance==0.2.33
plotly==5.18.0
webull==0.5.3
python-dotenv==1.0.0
gunicorn==21.2.0
requests==2.31.0
ib-insync==0.9.85
